trancoding fixes
