+++ src/api/queries/media/queries.ts
@@ -5,9 +5,6 @@ import {
 	useDownloadingDeviceProfileStore,
 	useStreamingDeviceProfileStore,
 } from '../../../stores/device-profile'
-import { usePlayerSettingsStore } from '../../../stores/settings/player'
-import { useUsageSettingsStore } from '../../../stores/settings/usage'
-import StreamingQuality from '../../../enums/audio-quality'
 import { SourceType } from '../../../types/JellifyTrack'
 import { ONE_DAY, queryClient } from '../../../constants/query-client'
 import { PlaybackInfoResponse } from '@jellyfin/sdk/lib/generated-client/models/playback-info-response'
@@ -20,18 +17,13 @@ export const MediaInfoQuery = (itemId: string | null | undefined, source: Source
 	const downloadingProfile = useDownloadingDeviceProfileStore.getState().deviceProfile
 	const profile = source === 'stream' ? streamingProfile : downloadingProfile
 
-	const streamingQuality = usePlayerSettingsStore.getState().streamingQuality
-	const downloadQuality = useUsageSettingsStore.getState().downloadQuality
-	const quality = source === 'stream' ? streamingQuality : downloadQuality
-	const isQualityLimited = quality !== StreamingQuality.Original
-
 	return {
 		queryKey: MediaInfoQueryKey({
 			api,
 			deviceProfile: profile,
 			itemId,
 		}),
-		queryFn: () => fetchMediaInfo(profile, itemId, isQualityLimited),
+		queryFn: () => fetchMediaInfo(profile, itemId),
 		enabled: Boolean(api && profile && itemId),
 		staleTime: ONE_DAY,
 	} as EnsureQueryDataOptions<PlaybackInfoResponse>
